week6_part1_app/requirements.txt
bigframes == 1.22.0
google-cloud-aiplatform == 1.70.0
google-cloud-pubsub == 2.26.0
google-cloud-storage == 2.18.2
gunicorn == 23.0.0
json_repair == 0.30.0
langchain_core == 0.3.10
langgraph == 0.2.35
mesop == 0.12.7
protobuf == 5.28.2
vertexai == 1.70.0
